Updates to pull out theming
