+++ src/app/features/applications/application/application-base.component.ts
@@ -9,7 +9,7 @@ import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Subscription } from 'rxjs/Subscription';
 import { RouterNav } from '../../../store/actions/router.actions';
 import { ApplicationEnvVarsService } from './application-tabs-base/tabs/build-tab/application-env-vars.service';
-import { GetAppSummaryAction } from '../../../store/actions/app-metadata.actions';
+import { GetAppInstancesAction, GetAppSummaryAction } from '../../../store/actions/app-metadata.actions';
 
 
 const applicationServiceFactory = (
@@ -80,6 +80,7 @@ export class ApplicationBaseComponent implements OnInit, OnDestroy {
     // Auto refresh
     this.entityServiceAppRefresh$ = this.entityService.poll(10000, this.autoRefreshString).do(() => {
       this.store.dispatch(new GetAppSummaryAction(appGuid, cfGuid));
+      this.store.dispatch(new GetAppInstancesAction(appGuid, cfGuid));
     }).subscribe();
 
     this.appSub$ = this.applicationService.app$.subscribe(app => {
