+++ src/app/features/applications/application/application-tabs-base/tabs/instances-tab/instances-tab.component.scss
