+++ src/app/features/applications/application/application-tabs-base/tabs/instances-tab/instances-tab.component.ts
@@ -1,12 +1,11 @@
-import { ApplicationMonitorService } from '../../application-monitor.service';
-import { CfAppInstancesConfigService } from '../../../../shared/list-configs/cf-app-instances-config.service';
 import { Component, OnInit } from '@angular/core';
-
 import { ActivatedRoute } from '@angular/router';
 
-import { ApplicationData, ApplicationService } from '../../application.service';
-import { CfAppEventsConfigService } from '../../../../shared/list-configs/cf-app-events-config.service';
-import { ListConfig } from '../../../../shared/components/list/list.component';
+import { ApplicationMonitorService } from '../../../../application-monitor.service';
+import { CfAppInstancesConfigService } from '../../../../../../shared/list-configs/cf-app-instances-config.service';
+import { ApplicationData, ApplicationService } from '../../../../application.service';
+import { CfAppEventsConfigService } from '../../../../../../shared/list-configs/cf-app-events-config.service';
+import { ListConfig } from '../../../../../../shared/components/list/list.component';
 
 @Component({
   selector: 'app-instances-tab',
@@ -15,7 +14,9 @@ import { ListConfig } from '../../../../shared/components/list/list.component';
   providers: [{
     provide: ListConfig,
     useClass: CfAppInstancesConfigService,
-  }]
+  },
+  ApplicationMonitorService
+]
 })
 export class InstancesTabComponent implements OnInit {
