+++ src/app/features/applications/applications.module.ts
@@ -17,7 +17,7 @@ import { ApplicationsRoutingModule } from './applications.routing';
 import { ApplicationTabsBaseComponent } from './application/application-tabs-base/application-tabs-base.component';
 import { DatePipe } from '@angular/common';
 import { EditApplicationComponent } from './edit-application/edit-application.component';
-import { InstancesTabComponent } from './application/instances-tab/instances-tab.component';
+import { InstancesTabComponent } from './application/application-tabs-base/tabs/instances-tab/instances-tab.component';
 import { ApplicationMonitorService } from './application-monitor.service';
