+++ src/app/shared/components/table/custom-cells/table-cell-usage/table-cell-usage.component.html
@@ -1,1 +1,1 @@
-<app-usage-gauge value="{{value(row)}}" valueText="{{label(row)}}"></app-usage-gauge>
+<app-usage-gauge warningAt="0.8" errorAt="0.9" value="{{value(row)}}" valueText="{{label(row)}}"></app-usage-gauge>
